big CSR refactor - protocol, ABC base, moving
