+++ coreblocks/priv/csr/shadow.py
@@ -10,9 +10,8 @@ from transactron.utils.dependencies import DependencyContext
 from transactron.lib import logging
 
 from coreblocks.func_blocks.csr.csr import CSRListKey
-from coreblocks.interface.layouts import CSRRegisterLayouts
 from coreblocks.params.genparams import GenParams
-from coreblocks.priv.csr.csr_register import CSRRegister
+from coreblocks.priv.csr.csr_register import CSRRegisterBase
 
 __all__ = ["ShadowCSR"]
 
@@ -20,7 +19,7 @@ __all__ = ["ShadowCSR"]
 log = logging.HardwareLogger("priv.csr.shadow")
 
 
-class ShadowCSR(CSRRegister):  # TODO: CSR register protocol
+class ShadowCSR(CSRRegisterBase):
     """CSR shadow register.
 
     Exposes an instruction-visible CSR number that reads/writes another CSR.
@@ -32,41 +31,30 @@ class ShadowCSR(CSRRegister):  # TODO: CSR register protocol
         self,
         csr_number: Optional[int],
         gen_params: GenParams,
-        shadowed: CSRRegister,
+        shadowed: CSRRegisterBase,
         *,
         mask: Optional[ValueLike | Method] = None,
         read_mask: Optional[ValueLike | Method] = None,
         write_mask: Optional[ValueLike | Method] = None,
         access_filter: Optional[Callable[[TModule, Value], ValueLike]] = None,
     ):
+        super().__init__(gen_params, csr_number, width=shadowed.width)
+
         if mask is not None:
             assert (
                 read_mask is None and write_mask is None
             ), "Cannot specify both full mask and separate read/write masks"
             read_mask = write_mask = mask
 
-        self.gen_params = gen_params
-        self.csr_number = csr_number
         self.shadowed = shadowed
-        self.width = shadowed.width
 
         full_mask = (1 << self.width) - 1
         self.read_mask: ValueLike | Method = full_mask if read_mask is None else read_mask
         self.write_mask: ValueLike | Method = full_mask if write_mask is None else write_mask
         self.access_filter = access_filter if access_filter is not None else (lambda _, __: C(1))
 
-        csr_layouts = gen_params.get(CSRRegisterLayouts)
-        self._fu_read = Method(o=csr_layouts._fu_read)
-        self._fu_write = Method(i=csr_layouts._fu_write)
-        self.value = Signal(self.width)
-
-        self.read = Method(o=csr_layouts.read)
-        self.read_comb = Method(o=csr_layouts.read)
-        self.write = Method(i=csr_layouts.write)
-        self.access_valid = Method(i=csr_layouts.access_valid_i, o=csr_layouts.access_valid_o)
-
         if csr_number is not None:
-            DependencyContext.get().add_dependency(CSRListKey(), self)
+            DependencyContext.get().add_dependency(CSRListKey(), (csr_number, self))
 
     def elaborate(self, platform):
         m = TModule()
@@ -126,8 +114,8 @@ class ShadowCSR(CSRRegister):  # TODO: CSR register protocol
                 "written": result.written,
             }
 
-        @def_method(m, self.access_valid)
+        @def_method(m, self._fu_access_valid)
         def _(priv_mode):
-            return {"valid": self.access_filter(m, priv_mode) & self.shadowed.access_valid(m, priv_mode).valid}
+            return {"valid": self.access_filter(m, priv_mode) & self.shadowed._fu_access_valid(m, priv_mode).valid}
 
         return m
